apps/api/requirements.txt
email-validator>=2.2.0,<3.0.0
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
alembic==1.16.5
psycopg[binary]==3.2.10
pydantic==2.11.9
pydantic-settings==2.11.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.20
cryptography==45.0.6
celery==5.5.3
redis==6.4.0
openai==1.107.0
tenacity==9.1.2
structlog==25.4.0
opentelemetry-api==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-instrumentation-fastapi==0.58b0
httpx==0.28.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
python-docx==1.1.2
reportlab==4.2.2
pypdf==5.4.0

pytest==8.4.2
pytest-asyncio==1.1.0
pytest-mock==3.15.1

ruff==0.13.0
black==25.1.0
mypy==1.18.2
types-python-jose==3.5.0.20250531
types-redis==4.6.0.20241004
